Asteroids falling on planet locations. - Scripts, prefabs, explosions!
This commit is the culmination of careful code organization sacrificed
upon the altar of the 10 day game jam time limit. While the early code
was likely too segmented in terms of DLL separation and abstraction
layers, we're now in the wild-west of everything in one assembly, and
apparently we're building a questionable amount of game content into
the UI layer.

Oops!

No time for regrets. Even this commit message is too much.